+++ src/browser/StorybookPage.ts
@@ -6,7 +6,7 @@ import type {
 } from 'storybook/internal/types';
 import pTimeout from 'p-timeout';
 import type {Page} from 'playwright';
-import dedent from 'ts-dedent';
+import {dedent} from 'ts-dedent';
 
 // Functions we pass to `page.evaluate` execute in a browser environment, and can access window.
 // eslint-disable-next-line no-var