+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- [breaking] Support Storybook 9
+
 ## 8.2.3 (2025-05-01)
 
 - [fix] update contents of package.json using `npm pkg fix` [#105](https://github.com/chanzuckerberg/axe-storybook-testing/pull/105)