+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - [breaking] Support Storybook 9
+- [breaking] Output ESM instead of CJS (needed to consume Storybook 9 types)
 
 ## 8.2.3 (2025-05-01)